@@ -4,7 +4,7 @@
 
   <groupId>no.paneon.oas.plantuml</groupId>
   <artifactId>pumloverlay</artifactId>
-  <version>6.6.11-SNAPSHOT</version>
+  <version>7.0.0-SNAPSHOT</version>
   <packaging>jar</packaging>
 
   <name>pumloverlay</name>
@@ -12,10 +12,10 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-	<version.java>11</version.java>
-	<java.version>11</java.version>
-	<maven.compiler.target>1.11</maven.compiler.target>
-	<maven.compiler.source>1.11</maven.compiler.source>
+	<version.java>21</version.java>
+	<java.version>21</java.version>
+	<maven.compiler.release>17</maven.compiler.release>
+
 	<git-commit-id-plugin.version>4.0.4</git-commit-id-plugin.version>
   </properties>
 
@@ -32,7 +32,7 @@
 	<dependency>
 		<groupId>no.paneon.api.diagram</groupId>
 		<artifactId>apidiagram</artifactId>
-		<version>6.6.11-SNAPSHOT</version>
+		<version>7.0.0-SNAPSHOT</version>
 	</dependency>
 
 	<!--dependency>
@@ -147,10 +147,6 @@
 			<groupId>org.apache.maven.plugins</groupId>
 			<artifactId>maven-compiler-plugin</artifactId>
 			<version>3.8.0</version>
-			<configuration>
-				<source>11</source>
-				<target>11</target>
-			</configuration>
 		</plugin>
 		
 		<plugin>
